# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HackSoft API",
        "description": "JSON-based API for social media application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8009/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/me": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user profile",
                "description": "Retrieve the currently authenticated user's profile with statistics",
                "operationId": "71adc4eb6e304863c1f5ea3852f29953",
                "responses": {
                    "200": {
                        "description": "User profile with statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "stats": {
                                            "properties": {
                                                "total_posts": {
                                                    "description": "Total number of posts by the user",
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total_likes": {
                                                    "description": "Total number of likes received on user's posts",
                                                    "type": "integer",
                                                    "example": 42
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update current user profile",
                "description": "Update profile information (name, description, and/or profile picture). Uses POST with _method=PATCH due to file upload requirements.",
                "operationId": "06de43c32f1b8208271e78dd459e9a3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "HTTP method override for file uploads",
                                        "type": "string",
                                        "enum": [
                                            "PATCH"
                                        ],
                                        "example": "PATCH"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "User's changed the name.",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "User's changed the description.",
                                        "nullable": true
                                    },
                                    "profile_picture": {
                                        "description": "Profile picture image file (max 2MB)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": "profile-pictures/image123.jpg",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Just a user.",
                        "nullable": true
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}